requirements.txt
torch==1.4.0
numpy==1.19.5
opencv-python==4.5.3.56
Pillow==8.4
scikit-image==0.19.3
matplotlib==3.5.2